Use a comparable number to identify the inputs
Since the two kinds of sieves have different inputs, the benchmark data
was hard to compare. One needs the size of the search space, and the
other just needs to know how many primes you want to gather.

This commit makes it so we can make one input struct that can be used to
identify both in a comparable metric. This way the graphs produced can
be more easily compared between the implementations.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>